companion/war-room/requirements.txt
google-genai>=1.0.0
sounddevice>=0.4.7
numpy>=1.26.0
